+++ js/gen/app.config.js
@@ -1,6 +1,6 @@
 app.config = {
 	env					: 'prod',
-	version				: '1.8.60', // if(session_data.user_id && session_data.user_id !== 'false')added to _processing_signup and added spinner for download community tracks
+	version				: '1.8.61', // set is_internet_connected to true in the _update_sisbot() in the env conditional for alpha to show Community for apple testing. Changed the api_url and web_url in the app.config in the alpha envs to point to https://webcenter.sisyphus-industries.com/, changing config #s
 	disconnect_timeout_to_stop_polling: 45000, // stop trying to find tables after 45 seconds
 	envs	: {
 		alpha: {	// loads local data only **5 tap on No Table Found Screen to launch shell app